refactor: create standalone template with environment-specific implementations
