+++ templates/standalone/config/application.conf
