+++ templates/standalone/config/clients.yaml
