+++ templates/standalone/config/clients.yaml.example
