+++ templates/standalone/docker-compose.test.yml
