+++ templates/standalone/package.json
@@ -1,0 +1,38 @@
+{
+  "name": "@o3co/auth-provider-standalone",
+  "version": "0.0.0",
+  "private": true,
+  "license": "Apache-2.0",
+  "type": "module",
+  "scripts": {
+    "build": "tsc",
+    "start": "node dist/app.mjs",
+    "debug": "tsx watch src/app.mts",
+    "test": "vitest run"
+  },
+  "imports": {
+    "#/*": "./src/*"
+  },
+  "dependencies": {
+    "@o3co/auth-provider-core": "workspace:*",
+    "@o3co/js.util.log": "^0.5.0",
+    "@o3co/ts.hocon": "^0.1.2",
+    "connect-redis": "^9.0.0",
+    "express": "^5.2.1",
+    "express-session": "^1.19.0",
+    "helmet": "^8.1.0",
+    "passport": "^0.7.0",
+    "passport-local": "^1.0.0",
+    "passport-oauth2-client-password": "^0.1.2",
+    "passport-google-oauth20": "^2.0.0",
+    "redis": "^5.10.0"
+  },
+  "devDependencies": {
+    "@types/express-session": "^1",
+    "@types/node": "^25.1.0",
+    "msw": "^2.12.14",
+    "pino": "^10.3.1",
+    "tsx": "^4.21.0",
+    "vitest": "^4.1.2"
+  }
+}
