+++ templates/standalone/tests/Makefile
