+++ templates/standalone/tests/did-grant.test.js
