+++ templates/standalone/tests/package.json
